requirements.txt
Flask==0.11.1
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.11
click==6.6
fuzzywuzzy==0.14.0
itsdangerous==0.24
python-Levenshtein==0.12.0
wsgiref==0.1.2
PyJWT==1.4.2
